move highmem routines into it's own file
