+++ kern/HighMem.h
@@ -1,0 +1,5 @@
+#ifndef __HIGHMEM_H__
+#define __HIGHMEM_H__
+void init_highmem();
+void term_highmem();
+#endif
